add family names to static site
this file allows us to add the common family names to the information panel. the current panel only documents latin names. 

at the moment, we don't make use of the the EOL IDs and URLs (they would make it possible to have a family information panel with imagery and links to species) 

I copied the file from the data pipeline repo