switch to larger png images
thank you @theswerd, looks great! I just switched the images to the larger versions Jess provided and switched their order. I think it's fine to leave out the link sharing for now.

it seems like there's a lot of extra space betwen the icons (it's 24px in the zeplin) and the section divider underneath. does the spacing there seem bigger to you too? I don't understand how it could be different from the rest of the sidebar, it looks consistent in the html file.