remove Zeplin and roadmap
Since Zeplin is only accessible by invitation (I think that's just a function of Jess' account), let's either not link to it or specify that someone can request access to it. https://public-tree-map.github.io/  is set up as a roadmap, so for now I'd prefer to build on what's there. clarifying the project's current status and defining more short term goals, like a landing page, is a great idea